Rename donation_dialog.jsx -> donationDialog.jsx
